MOved promooImageextractor and buildimage functions to utilty module
