+++ src/app/hooks/useUASButton/index.ts
@@ -8,11 +8,10 @@ import {
   buildGlobalId,
   FAVOURITES_CONFIG,
   createFavouritesPayload,
+  extractPromoImageFromArticleData,
+  buildPromoImageUrl,
 } from '#app/lib/uasApi/uasUtility';
-import buildIChefURL from '#app/lib/utilities/ichefURL';
-import extractPromoImage from '#app/lib/utilities/extractPromoImage';
 import type { SaveArticleButtonProps } from '#app/components/SaveArticleButton';
-import type { OptimoRawImageBlock } from '#app/models/types/optimo';
 import useToggle from '../useToggle';
 
 /** A hook that fetches an article's saved status and controls showing the save UAS button
@@ -31,38 +30,6 @@ interface UseUASButtonReturn {
   error: Error | null;
   handleSaveAction: (action: UASAction) => Promise<void>;
 }
-
-const extractPromoImageFromArticleData = (
-  articlePageData?: SaveArticleButtonProps['articlePageData'],
-) => {
-  const promoImageBlocks =
-    articlePageData?.promo?.images?.defaultPromoImage?.blocks ?? [];
-
-  const { altText, rawBlock } = extractPromoImage(promoImageBlocks);
-
-  return {
-    altText,
-    promoImageRawBlock: rawBlock,
-  };
-};
-
-const buildPromoImageUrl = (promoImageObj?: {
-  altText: string;
-  promoImageRawBlock?: OptimoRawImageBlock;
-}): string => {
-  if (
-    !promoImageObj?.promoImageRawBlock?.model?.locator ||
-    !promoImageObj?.promoImageRawBlock?.model?.originCode
-  ) {
-    return '';
-  }
-
-  return buildIChefURL({
-    originCode: promoImageObj.promoImageRawBlock.model.originCode,
-    locator: promoImageObj.promoImageRawBlock.model.locator,
-    resolution: 320,
-  });
-};
 
 const useUASButton = ({
   articleId,
